Removing list command deprecation. Removing search command
